Add files via upload
Direct link - http://localhost:8888/notebooks/Desktop/REAL%20ESTATE%20PROJECT/REAL%20ESTATE.ipynb
This is the data of a random Real Estate company, I just visualize this data and analyze Property Prices variance, things on which it depends, and graphs of the data variance. 
You will get this source code on this REAL ESTATE - checkpoint.ipynb folder, I uploaded the rest of the other folders which are linked with this repository.